+++ src/helpers/entities/vacuum/vacuum.js
@@ -1,4 +1,4 @@
-import {Sensor} from '../sensor/sensor';
+import {BatterySensor, Sensor} from '../sensor/sensor';
 import {
   VACUUM_ACTIONS_ICONS,
   VACUUM_ACTION_PAUSE,
@@ -15,24 +15,43 @@ export class VacuumEntity {
     this.entity_id = config.entity;
     let entity = hass.states[config.entity];
     this.state = entity.state;
-    this.battery_icon = entity.attributes.battery_icon;
-    this.battery_level = entity.attributes.battery_level;
-    this.fan_speed = entity.attributes.fan_speed;
+    this._sensors = config.sensors;
+    this._fan_speed = entity.attributes.fan_speed;
     this.friendly_name = entity.attributes.friendly_name;
 
-    this._sensors = config.sensors;
     this._service = config.shortcuts.service
       ? config.shortcuts.service.split('.')[0]
       : null;
     this._service_action = config.shortcuts.service
       ? config.shortcuts.service.split('.')[1]
       : null;
+    this._command = config.shortcuts.command ? config.shortcuts.command : null;
     this.shortcuts = config.shortcuts.description;
     this._actions = {};
     this._actions[VACUUM_ACTION_START] = config.start;
     this._actions[VACUUM_ACTION_PAUSE] = config.pause;
     this._actions[VACUUM_ACTION_STOP] = config.stop;
     this._actions[VACUUM_ACTION_RETURN_TO_BASE] = config.return_to_base;
+  }
+
+  get fan_speed() {
+    return this._fan_speed ? this._fan_speed : null;
+  }
+
+  get name() {
+    return this.friendly_name ? this.friendly_name : this.entity_id;
+  }
+
+  get battery() {
+    return this._sensors['battery']
+      ? new BatterySensor(this._sensors['battery'], this._hass)
+      : null;
+  }
+
+  get mop_intensite() {
+    return !this._sensors['mop_intensite']
+      ? null
+      : new Sensor(this._sensors['mop_intensite'], this._hass);
   }
 
   get current_clean_area() {
@@ -47,31 +66,16 @@ export class VacuumEntity {
       : new Sensor(this._sensors['current_clean_duration'], this._hass);
   }
 
-  get last_clean_area() {
-    return !this._sensors['last_clean_area']
+  get map() {
+    return !this._sensors['map']
       ? null
-      : new Sensor(this._sensors['last_clean_area'], this._hass);
-  }
-
-  get last_clean_duration() {
-    return !this._sensors['last_clean_duration']
-      ? null
-      : new Sensor(this._sensors['last_clean_duration'], this._hass);
-  }
-
-  get camera() {
-    return !this._sensors['camera']
-      ? null
-      : new Sensor(this._sensors['camera'], this._hass);
+      : new Sensor(this._sensors['map'], this._hass);
   }
 
   get sensors() {
-    return [
-      this.current_clean_area,
-      this.current_clean_duration,
-      this.last_clean_area,
-      this.last_clean_duration,
-    ].filter((n) => n);
+    return [this.current_clean_area, this.current_clean_duration].filter(
+      (n) => n
+    );
   }
 
   get actions() {
@@ -97,12 +101,11 @@ export class VacuumEntity {
     return {
       entity_id: this.entity_id,
       attributes: {
-        battery_icon: this.battery_icon,
-        battery_level: this.battery_level,
-        fan_speed: this.fan_speed,
         friendly_name: this.friendly_name,
+        battery: this.battery,
+        fan_speed: this.fan_speed,
         icon: this.icon,
-        map_url: this.camera ? this.camera.attributes.entity_picture : null,
+        map_url: this.map,
       },
       state: this.state,
     };
@@ -115,11 +118,15 @@ export class VacuumEntity {
   }
 
   _callShortcutService(id) {
+    if (!this._service) throw new Error('Service to call is not defined');
+    if (!this._command)
+      throw new Error('Service command to call is not defined');
     const shortcuts = this.shortcuts[id];
     if (!shortcuts.segments)
       throw new Error('Segment list for ' + shortcuts.name + 'is empty');
     return this._hass.callService(this._service, this._service_action, {
       entity_id: [this.entity_id],
+      command: this._command,
       segments: shortcuts.segments,
     });
   }